+++ lib/brando/sites/identity.ex
@@ -11,7 +11,7 @@ defmodule Brando.Sites.Identity do
   use Gettext, backend: Brando.Gettext
 
   trait :timestamped
-  trait Brando.Trait.Translatable, alternates: false
+  trait :translatable, alternates: false
 
   identifier false
   persist_identifier false